Refactored storage handling.
Now uses a storage cache that gets updated when other tabs write to storage.
And similalry notifies other tabs when it writes to storage.

Still need to invalidate and redraw gui when storage is mutated by remote.
 This might take some work.
Also seems to notify itself as well, not only other tabs,
 which shouldn't, but might, cause cyclical issues?
Plenty of TODO's left to handle.
There is an attempt at using proxies to hiding the fact that it is a wrapping api for storage,
 allowing user to handle the storage object as any other and have it automatically sync cache with storage.
 Additionally some work on a recursive proxy version,
  so nested objects also propagate mutation events to let storage know it needs ot be synced.
 This is currently having some issues and needs some work.